Add DatingGameTestbench used for development
Once Unity compiles the code, it finds things in the "Editor" directory
to load into the Unity Editor itself. This let me extend the
functionality of the Editor through the DatingGameTestbench menu to run
some of our game code without "running the game" or making test menus in
the game itself.